save favorites is work after refactoring
