+++ core/src/main/java/com/amk/core/repository/RepositoryCompany.kt
@@ -4,9 +4,9 @@ import com.amk.core.entity.Company
 
 interface RepositoryCompany {
 
-    suspend fun CreateListOneDayYesterday(): List<Company>
+    suspend fun сreateListOneDayYesterday(): List<Company>
 
-    suspend fun CreateListOneDayHalfYear(): List<Company>
+    suspend fun сreateListOneDayHalfYear(): List<Company>
 
     suspend fun addFavoriteCompany(secId: String)
