+++ core/src/main/java/com/amk/core/repository/RepositoryCompanyImpl.kt
@@ -5,6 +5,7 @@ import android.content.Context.MODE_PRIVATE
 import com.amk.core.entity.BaseCashCompany
 import com.amk.core.entity.Company
 import com.amk.core.entity.EntityCompany
+import com.amk.core.entity.FavoriteCompany
 import com.amk.core.interactors.CompanyFactory
 import com.amk.core.utils.DATA_LOAD
 import com.amk.core.utils.convertToDate
@@ -21,7 +22,7 @@ class RepositoryCompanyImpl(
     private val lastData = sharedPref.getString(DATA_LOAD, "")
     private val todayData = Date().convertToString()
 
-    override suspend fun CreateListOneDayYesterday(): List<Company> {
+    override suspend fun сreateListOneDayYesterday(): List<Company> {
         val listFavorite = cacheRepository.getFavoriteCompanies().map { it.secId }.toList()
         val dataGetReadyIsShow = mutableListOf<Company>()
 
@@ -37,7 +38,7 @@ class RepositoryCompanyImpl(
         return dataGetReadyIsShow
     }
 
-    override suspend fun CreateListOneDayHalfYear(): List<Company> {
+    override suspend fun сreateListOneDayHalfYear(): List<Company> {
         val listFavorite = cacheRepository.getFavoriteCompanies().map { it.secId }.toList()
         val dataGetReadyIsShow = mutableListOf<Company>()
